improvements to BlockSource and various removals of inactive & obsolete code
